New PocoBuilder parses its first resource succesfully.
