+++ src/Hl7.Fhir.Base/ElementModel/PocoBuilder.cs
@@ -79,9 +79,6 @@ namespace Hl7.Fhir.Serialization
                 ErrorMode = _settings.IgnoreUnknownMembers ?
                     TypedElementSettings.TypeErrorMode.Ignore
                     : TypedElementSettings.TypeErrorMode.Report,
-#pragma warning disable CS0618 // Type or member is obsolete
-                TruncateDateTimeToDate = _settings.TruncateDateTimeToDate
-#pragma warning restore CS0618 // Type or member is obsolete
             };
 
             string dataType;
@@ -130,4 +127,3 @@ namespace Hl7.Fhir.Serialization
         }
     }
 }
-
